LayoutTests/fast/css/cssom-remove-shorthand-property-expected.txt

Tests that shorthand properties can be removed via CSSOM.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canSetProperty('font') is true
PASS canRemoveProperty('font') is true
PASS canSetProperty('border-top') is true
PASS canRemoveProperty('border-top') is true
PASS canSetProperty('border-right') is true
PASS canRemoveProperty('border-right') is true
PASS canSetProperty('border-bottom') is true
PASS canRemoveProperty('border-bottom') is true
PASS canSetProperty('border-left') is true
PASS canRemoveProperty('border-left') is true
PASS canSetProperty('border') is true
PASS canRemoveProperty('border') is true
PASS canSetProperty('border-color') is true
PASS canRemoveProperty('border-color') is true
PASS canSetProperty('border-style') is true
PASS canRemoveProperty('border-style') is true
PASS canSetProperty('border-width') is true
PASS canRemoveProperty('border-width') is true
PASS canSetProperty('background-position') is true
PASS canRemoveProperty('background-position') is true
PASS canSetProperty('background-repeat') is true
PASS canRemoveProperty('background-repeat') is true
PASS canSetProperty('border-spacing') is true
PASS canRemoveProperty('border-spacing') is true
PASS canSetProperty('list-style') is true
PASS canRemoveProperty('list-style') is true
PASS canSetProperty('margin') is true
PASS canRemoveProperty('margin') is true
PASS canSetProperty('outline') is true
PASS canRemoveProperty('outline') is true
PASS canSetProperty('padding') is true
PASS canRemoveProperty('padding') is true
PASS canSetProperty('background') is true
PASS canRemoveProperty('background') is true
PASS canSetProperty('overflow') is true
PASS canRemoveProperty('overflow') is true
PASS canSetProperty('border-radius') is true
PASS canRemoveProperty('border-radius') is true
PASS successfullyParsed is true

TEST COMPLETE
